backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
openai==0.28.1
faiss-cpu==1.7.4
numpy==1.26.0
pydantic==2.4.2